(feat) Lots of reconfiguration - too big of a commit. Sockets can now be used to transmit events between multiple clients and the server
