+++ client/actionCreators/chat.js
@@ -1,5 +1,7 @@
 import fetch from 'isomorphic-fetch';
 
+import { socket } from '../lib/socketSetup';
+
 export function fetchMessages(rideId) {
   return (dispatch) => {
     dispatch(requestMessages());
@@ -27,9 +29,19 @@ function requestMessagesError(error) {
   return { type: 'REQUEST_MESSAGES_ERROR', entry: error, };
 };
 
+// message: { userId, time, text }
 export function addMessage(message) {
-  return {
-    type: 'ADD_MESSAGE',
-    entry: message,
-  };
+  console.log('adding a message!');
+  return { type: 'ADD_MESSAGE', entry: message, };
 };
+
+// What does this message look like?
+export function submitMessage(message) {
+  console.log('submitting message!');
+  socket.emit('send_message', message);
+
+  // return addMessage creator if we want to automatically add the message
+  // instead of waiting for the message to be retransmitted by the server.
+  // return addMessage(message);
+  return { type: 'Not an action', entry: message };
+}
